LPS-155662 New aproach to home folder structure
