+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/home/components/ProjectsList/components/ProjectCard/components/CardTitleDescription/index .js
@@ -9,4 +9,4 @@
  * distribution rights of the Software.
  */
 
-export {default} from './StatusDescription';
+export {default} from './CardTitleDescription';
