+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/home/components/SearchHeader/index.js
@@ -9,38 +9,4 @@
  * distribution rights of the Software.
  */
 
-import {useState} from 'react';
-import i18n from '../../../../common/I18n';
-import Skeleton from '../../../../common/components/Skeleton';
-import SearchBar from './components/SearchBar/SearchBar';
-
-const SearchHeader = ({count, loading, onSearchSubmit}) => {
-	const [searchTerm, setSearchTerm] = useState('');
-
-	const getCounter = () => {
-		return `${count} ${
-			searchTerm
-				? i18n.pluralize(count, 'result')
-				: i18n.pluralize(count, 'project')
-		}`;
-	};
-
-	return (
-		<div className="align-items-center d-flex justify-content-between mb-4">
-			<SearchBar
-				onSearchSubmit={(term) => {
-					setSearchTerm(term);
-					onSearchSubmit(term);
-				}}
-			/>
-
-			{loading ? (
-				<Skeleton height={22} width={85} />
-			) : (
-				<h5 className="m-0 text-neutral-7">{getCounter()}</h5>
-			)}
-		</div>
-	);
-};
-
-export default SearchHeader;
+export {default} from './SearchHeader';
