+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/home/index.js
@@ -9,56 +9,4 @@
  * distribution rights of the Software.
  */
 
-import classNames from 'classnames';
-import ProjectList from './components/ProjectsList';
-import SearchHeader from './components/SearchHeader';
-import useHasManyProjects from './hooks/useHasManyProjects';
-import useKoroneikiAccounts from './hooks/useKoroneikiAccounts';
-import './app.scss';
-
-const Home = () => {
-	const {data, fetchMore, fetching, loading, search} = useKoroneikiAccounts();
-	const koroneikiAccounts = data?.c?.koroneikiAccounts;
-
-	const hasManyProjects = useHasManyProjects(koroneikiAccounts);
-
-	return (
-		<div
-			className={classNames({
-				'cp-project-cards-container': !hasManyProjects,
-				'mx-auto cp-project-cards-container-sm': hasManyProjects,
-			})}
-		>
-			<div
-				className={classNames({
-					'd-flex flex-column w-100': hasManyProjects,
-					'ml-3': !hasManyProjects,
-				})}
-			>
-				{hasManyProjects && (
-					<SearchHeader
-						count={koroneikiAccounts?.totalCount}
-						loading={loading}
-						onSearchSubmit={(term) => search(term)}
-					/>
-				)}
-
-				<ProjectList
-					fetching={fetching}
-					hasManyProjects={hasManyProjects}
-					koroneikiAccounts={koroneikiAccounts}
-					loading={loading}
-					onIntersect={(currentPage) =>
-						fetchMore({
-							variables: {
-								page: currentPage + 1,
-							},
-						})
-					}
-				/>
-			</div>
-		</div>
-	);
-};
-
-export default Home;
+export {default} from './Home';
